requirements.txt
streamlit
langchain
langgraph
langchain-community
langchain-huggingface
langchain-groq
chromadb
pypdf2
speechrecognition
pydub
reportlab
python-dotenv
scikit-learn
pandas
numpy